# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mi API",
        "description": "Documentación de la API para el sistema UsersPosts",
        "contact": {
            "email": "contacto@miapp.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/posts": {
            "get": {
                "tags": [
                    "posts"
                ],
                "summary": "Lista de los posts creados por el usuario autenticado",
                "operationId": "4186b281454d4cf4c67a3254c731218a",
                "responses": {
                    "200": {
                        "description": "Lista de posts paginados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "posts": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "title": {
                                                                "type": "string",
                                                                "example": "Titulo de prueba"
                                                            },
                                                            "description": {
                                                                "type": "string",
                                                                "example": "Descripcion de prueba"
                                                            },
                                                            "content": {
                                                                "type": "string",
                                                                "example": "contenido de prueba de minimo 50 caracteres ..."
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 100
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "El usuario autenticado no tiene posts creados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Posts not founded for this user"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "posts"
                ],
                "summary": "Crea un nuevo post y lo vincula con el usuario autenticado como su creador.",
                "operationId": "c525736cca34af3402c290c5a680dbde",
                "requestBody": {
                    "description": "Datos para crear un post",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post creado exitosamente. Retorna los datos del post",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "title",
                                        "description",
                                        "content"
                                    ],
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "post": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Titulo Ejemplo"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Descripción de ejemplo"
                                                },
                                                "content": {
                                                    "type": "string",
                                                    "example": "Contenido de ejemplo minimo 50 caracteres ..."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validación de datos"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticación usando un token JWT",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}